FEAT(#205): ADD CREATE AND PROYECT FORM
I have changed some things in the project model that I realized were missing, and I have also tried to cover all the errors so that they appear in the form. I have not managed to get the messages in Spanish to appear, but since a specific task is going to be created for these messages, I have left it that way for now. If you know how to modify that and you want me to change it here, please tell me. If you see anything that you see inappropriate and it is necessary to change/modify it, I say the same. To test this pr you just have to go to the project/create route and try the different attributes. I have redirected to the main page because there is no project list view.