FEAT(#211): CREATE PROYECT DELETE FUNCTION
To test this pr, it is necessary to create an administrator, create a project from the admin section, and then insert the url project/id/delete, then check that it has been deleted in the admin panel.
The model has flaws, but I have seen that in another pr that I have fixed and it is in the process of review, in this pr I have only concentrated on deletion.